// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  actions: typeof actions;
  auth: typeof auth;
  bookmarks: typeof bookmarks;
  http: typeof http;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;